SUKU files moved, files deleted
