+++ src/app/main/panels/configuration/ConfigParameters.svelte
@@ -4,7 +4,7 @@
 
   import { get } from 'svelte/store';
   import { runtime, engine, logger, user_input, controlElementClipboard} from '../../../runtime/runtime.store.js';
-  import { configManagement } from "../../../runtime/config-manager.store.js";
+  import { configManagement } from "./Configuration.store.js";
   import { onDestroy, onMount } from "svelte";
   import _utils from "../../../runtime/_utils.js";
